server/requirements.txt
fastapi[standard]
firebase-admin
requests
python-dotenv
uvicorn[standard]
numpy
